Fix integer overflow in adeclick filter size computation

+++ trunk/libs/libav/ffmpeg-5.0.1/libavfilter/af_adeclick.c
@@ -340,7 +340,7 @@ static int interpolation(DeclickChannel *c, const double *src, int ar_order,
     double *vector, *matrix;
     int i, j;
 
-    av_fast_malloc(&c->matrix, &c->matrix_size, nb_errors * nb_errors * sizeof(*c->matrix));
+    av_fast_malloc(&c->matrix, &c->matrix_size, (size_t)nb_errors * nb_errors * sizeof(*c->matrix));
     matrix = c->matrix;
     if (!matrix)
         return AVERROR(ENOMEM);
